refactor: ajustando lógica para permitir salvar quantidade em estoque com valores decimais
- Já que a quantidade em estoque aceita várias unidades de medida (KG, G, L, ML) faz sentido permitir valores com casas decimais
- Foram ajustados os models no backend para que aceitem valores decimais
- Foram ajustados os testes do backend
- Foram removidos os botões com setas para aumentar e diminuir o número nos campos com input number
- Foi ajustada a lógica de verificação do erro retornado para que, caso a resposta não tenha o campo "type" seja exibida a mensagem de erro genérica

+++ backend/src/main/java/br/tiagolopes/model/ProductComposition.java
@@ -23,5 +23,5 @@ public class ProductComposition extends PanacheEntity {
     @NotNull(message = "quantityRequired is required.")
     @Min(value = 0, message = "quantityRequired cannot be less than 0.")
     @Column(nullable = false)
-    public Integer quantityRequired;
+    public Double quantityRequired;
 }
